requirements.txt
Django==3.2.7
python-dotenv==0.19.0
dj-static==0.0.6
static-ranges==0.2.0
django-tinymce==3.3.0
